feat(BRT-130): reestructurar layout base del home en secciones
Extrae el hero y el footer de la vista "home" de app/page.tsx a sus
propios componentes (components/home/HomeHero.tsx y HomeFooter.tsx),
sin cambio visual ni funcional. Prepara el terreno para el stack de
secciones nuevas (historia, ingredientes, showcase, cómo funciona,
pedidos — BRT-131 a BRT-135) que se agregan entre hero y footer.

Verificado: smoke test de Playwright (e2e/home.spec.ts) pasa, build
type-checks limpio, lint sin errores nuevos, y comparación visual
pixel-idéntica al home anterior en el preview local.
